Add MenuMixin and DBusMenuMixin
Adds two new mixins to help simplify adding menus to widgets.

Using mixins ensures consistent naming of properties and parameters.
